+++ application/features/bootstrap/NotificationContext.php
@@ -58,4 +58,44 @@ class NotificationContext extends SyncContext
 
         $this->makeFakeIdStoreWithUser($tempIdStoreUserInfo);
     }
+
+    /**
+     * @Then the email subject contains :subject
+     */
+    public function theEmailSubjectContains($subject)
+    {
+        Assert::assertNotEmpty($this->notifier->findEmailBySubject($subject));
+    }
+
+    /**
+     * @Then an email is not sent
+     */
+    public function anEmailIsNotSent()
+    {
+        Assert::assertEmpty($this->notifier->emailsSent);
+    }
+
+    /**
+     * @Then an email with subject :subject is not sent
+     */
+    public function anEmailWithSubjectIsNotSent($subject)
+    {
+        Assert::assertEmpty($this->notifier->findEmailBySubject($subject));
+    }
+
+    /**
+     * @Then a :subject email is sent to the user's HR contact
+     */
+    public function aEmailIsSentToTheUsersHrContact($subject)
+    {
+        $email = $this->notifier->findEmailBySubject($subject);
+        Assert::assertNotEmpty($email, "No email was found with the subject: " . $subject);
+
+        $user = $this->idStore->getActiveUser($this->tempEmployeeId);
+        Assert::assertContains(
+            $user->getHRContactEmail(),
+            $email['to_address'],
+            "Email was not sent to " . $user->getHRContactEmail()
+        );
+    }
 }